2.0.7
1. Delete github link group
2. Rollback StackUtils.java to 2.0.2
3. Feat: Lazy mode for advanced line transfers
4. Fix: The big lag in line transfers
5. Rewrite Advanced Line Transfer (Grabber)
6. Version Update

+++ src/main/java/io/github/sefiraat/networks/Networks.java
@@ -51,7 +51,7 @@ public class Networks extends JavaPlugin implements SlimefunAddon {
     private ConfigManager configManager;
     private ListenerManager listenerManager;
     private SupportedPluginManager supportedPluginManager;
-    private int slimefunTickCount;
+    private long slimefunTickCount;
 
 
     public Networks() {
@@ -93,7 +93,7 @@ public class Networks extends JavaPlugin implements SlimefunAddon {
         return Networks.getInstance().listenerManager;
     }
 
-    public static int getSlimefunTickCount() {
+    public static long getSlimefunTickCount() {
         return getInstance().slimefunTickCount;
     }
